requirements.txt
feedparser
requests
beautifulsoup4
readability-lxml
python-dateutil
faiss-cpu
langchain-community
langchain-core
langchain-text-splitters
langchain-huggingface
sentence-transformers
